改进 ChatHistory 为 ChatMessage 增加 Redis

+++ internal/entity/Base.go
@@ -1,7 +1,7 @@
 package entity
 
 type Base struct {
-	ID        int    `xorm:"id int(11) notnull pk autoincr" json:"id"`
+	ID        int64  `xorm:"id int(11) notnull pk autoincr" json:"id"`
 	CreatedAt string `xorm:"created timestamp" json:"created_at"`
 	UpdatedAt string `xorm:"updated timestamp" json:"updated_at"`
 }
